Split sidebar observation helpers
