+++ Sources/Workspace.swift
@@ -9785,44 +9785,6 @@ struct SidebarGitBranchState: Equatable {
     let isDirty: Bool
 }
 
-private struct SidebarPanelObservationState: Equatable {
-    let panelIds: [UUID]
-
-    init(panels: [UUID: any Panel]) {
-        panelIds = panels.keys.sorted { $0.uuidString < $1.uuidString }
-    }
-}
-
-private struct SidebarImmediateObservationState: Equatable {
-    let title: String
-    let customDescription: String?
-    let isPinned: Bool
-    let customColor: String?
-    let latestConversationMessage: String?
-    let latestSubmittedMessage: String?
-    let latestSubmittedAt: Date?
-}
-
-private struct SidebarObservationState: Equatable {
-    let currentDirectory: String
-    let extensionSidebarProjectRootPath: String?
-    let panels: SidebarPanelObservationState
-    let panelDirectories: [UUID: String]
-    let statusEntries: [String: SidebarStatusEntry]
-    let metadataBlocks: [String: SidebarMetadataBlock]
-    let logEntries: [SidebarLogEntry]
-    let progress: SidebarProgressState?
-    let gitBranch: SidebarGitBranchState?
-    let panelGitBranches: [UUID: SidebarGitBranchState]
-    let pullRequest: SidebarPullRequestState?
-    let panelPullRequests: [UUID: SidebarPullRequestState]
-    let remoteConfiguration: WorkspaceRemoteConfiguration?
-    let remoteConnectionState: WorkspaceRemoteConnectionState
-    let remoteConnectionDetail: String?
-    let activeRemoteTerminalSessionCount: Int
-    let listeningPorts: [Int]
-}
-
 enum WorkspaceRemoteConnectionState: String, Equatable {
     case disconnected
     case connecting
@@ -10685,99 +10647,8 @@ final class Workspace: Identifiable, ObservableObject {
     // Sidebar rows cache snapshots, so observation must begin with the current
     // workspace state. Build state publishers from @Published current values
     // instead of dropping the first value and repairing timing with a Void event.
-    lazy var sidebarImmediateObservationPublisher: AnyPublisher<Void, Never> = {
-        let workspaceFields = Publishers.CombineLatest4(
-            $title,
-            $customDescription,
-            $isPinned,
-            $customColor
-        )
-        let conversationFields = Publishers.CombineLatest3(
-            $latestConversationMessage,
-            $latestSubmittedMessage,
-            $latestSubmittedAt
-        )
-
-        return workspaceFields
-            .combineLatest(conversationFields)
-            .map { workspaceFields, conversationFields in
-                SidebarImmediateObservationState(
-                    title: workspaceFields.0,
-                    customDescription: workspaceFields.1,
-                    isPinned: workspaceFields.2,
-                    customColor: workspaceFields.3,
-                    latestConversationMessage: conversationFields.0,
-                    latestSubmittedMessage: conversationFields.1,
-                    latestSubmittedAt: conversationFields.2
-                )
-            }
-            .removeDuplicates()
-            .map { _ in () }
-            .eraseToAnyPublisher()
-    }()
-
-    lazy var sidebarObservationPublisher: AnyPublisher<Void, Never> = {
-        let workspaceFields = Publishers.CombineLatest4(
-            $currentDirectory,
-            $extensionSidebarProjectRootPath,
-            $panels.map(SidebarPanelObservationState.init),
-            $panelDirectories
-        )
-        let metadataFields = Publishers.CombineLatest4(
-            $statusEntries,
-            $metadataBlocks,
-            $logEntries,
-            $progress
-        )
-        let gitFields = Publishers.CombineLatest4(
-            $gitBranch,
-            $panelGitBranches,
-            $pullRequest,
-            $panelPullRequests
-        )
-        let remoteFields = Publishers.CombineLatest4(
-            $remoteConfiguration,
-            $remoteConnectionState,
-            $remoteConnectionDetail,
-            $activeRemoteTerminalSessionCount
-        )
-
-        return Publishers.CombineLatest4(
-            workspaceFields,
-            metadataFields,
-            gitFields,
-            remoteFields
-        )
-        .combineLatest($listeningPorts)
-        .map { groupedFields, listeningPorts in
-            let workspaceFields = groupedFields.0
-            let metadataFields = groupedFields.1
-            let gitFields = groupedFields.2
-            let remoteFields = groupedFields.3
-            return SidebarObservationState(
-                currentDirectory: workspaceFields.0,
-                extensionSidebarProjectRootPath: workspaceFields.1,
-                panels: workspaceFields.2,
-                panelDirectories: workspaceFields.3,
-                statusEntries: metadataFields.0,
-                metadataBlocks: metadataFields.1,
-                logEntries: metadataFields.2,
-                progress: metadataFields.3,
-                gitBranch: gitFields.0,
-                panelGitBranches: gitFields.1,
-                pullRequest: gitFields.2,
-                panelPullRequests: gitFields.3,
-                remoteConfiguration: remoteFields.0,
-                remoteConnectionState: remoteFields.1,
-                remoteConnectionDetail: remoteFields.2,
-                activeRemoteTerminalSessionCount: remoteFields.3,
-                listeningPorts: listeningPorts
-            )
-        }
-        .removeDuplicates()
-        .map { _ in () }
-        .eraseToAnyPublisher()
-    }()
+    lazy var sidebarImmediateObservationPublisher: AnyPublisher<Void, Never> = makeSidebarImmediateObservationPublisher()
+    lazy var sidebarObservationPublisher: AnyPublisher<Void, Never> = makeSidebarObservationPublisher()
 
     private func scheduleExtensionSidebarProjectRootRefresh(for directory: String) {
         extensionSidebarProjectRootRefreshID &+= 1
