+++ cmuxTests/WorkspaceUnitTests.swift
@@ -6421,66 +6421,6 @@ final class WorkspacePanelGitBranchTests: XCTestCase {
         )
     }
 
-    func testSidebarObservationPublisherEmitsForLateStatusSubscriber() {
-        let workspace = Workspace()
-        workspace.statusEntries["test_probe"] = SidebarStatusEntry(
-            key: "test_probe",
-            value: "VISIBLE?",
-            icon: "star.fill",
-            color: "#FF0000",
-            priority: 200
-        )
-
-        var publishCount = 0
-        let cancellable = workspace.sidebarObservationPublisher.sink {
-            publishCount += 1
-        }
-        defer { cancellable.cancel() }
-
-        XCTAssertGreaterThan(
-            publishCount,
-            0,
-            "A sidebar row that subscribes after status metadata already exists must still refresh from the current workspace state."
-        )
-    }
-
-    func testSidebarImmediateObservationPublisherEmitsForLateTitleSubscriber() {
-        let workspace = Workspace()
-        workspace.title = "Restored Workspace"
-
-        var publishCount = 0
-        let cancellable = workspace.sidebarImmediateObservationPublisher.sink {
-            publishCount += 1
-        }
-        defer { cancellable.cancel() }
-
-        XCTAssertGreaterThan(
-            publishCount,
-            0,
-            "A sidebar row that subscribes after immediate workspace fields already exist must still refresh from the current workspace state."
-        )
-    }
-
-    func testSidebarObservationPublisherIgnoresRemoteHeartbeatOnlyChanges() {
-        let workspace = Workspace()
-
-        var publishCount = 0
-        let cancellable = workspace.sidebarObservationPublisher.sink {
-            publishCount += 1
-        }
-        defer { cancellable.cancel() }
-        publishCount = 0
-
-        workspace.remoteHeartbeatCount = 1
-        workspace.remoteLastHeartbeatAt = Date()
-
-        XCTAssertEqual(
-            publishCount,
-            0,
-            "Expected non-visible remote heartbeat updates to avoid invalidating sidebar rows"
-        )
-    }
-
     @MainActor
     func testSidebarPullRequestsTrackFocusedPanelOnly() {
         let workspace = Workspace()
